Moved RenderAction class to its new folder
