+++ src/scene/composition/layer-composition.js
@@ -10,7 +10,7 @@ import {
     LIGHTTYPE_DIRECTIONAL, LIGHTTYPE_OMNI, LIGHTTYPE_SPOT
 } from '../constants.js';
 
-import { RenderAction } from '../render-action.js';
+import { RenderAction } from './render-action.js';
 import { WorldClusters } from '../world-clusters.js';
 import { LightCompositionData } from './light-composition-data.js';
